requirements.txt
Django==2.2.1
django-extensions==2.1.7
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
psycopg2==2.8.2
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
